expose lib to be tested
