+++ src/main.rs
@@ -4,10 +4,12 @@ extern crate cached;
 
 use warp::Filter;
 use std::convert::Infallible;
+use grekko::start_server;
+use std::net::{SocketAddrV4, Ipv4Addr, IpAddr, SocketAddr};
 
-mod request;
+pub mod request;
 mod vrp;
-mod geocoding;
+pub mod geocoding;
 
 // cached_key! {
 //     CACHE: SizedCache<String, Reader<File>> = SizedCache::with_size(1);
@@ -20,37 +22,6 @@ mod geocoding;
 
 #[tokio::main]
 async fn main() {
-
-    let forward_geocoding = warp::path!("geocoding" / "forward" / String)
-        .and_then(receive_and_search_coordinates);
-
-    let reverse_geocoding = warp::path!("geocoding" / "reverse" / f64 / f64)
-        .and_then(receive_and_search_postcode);
-
-    let trip = warp::post()
-        .and(warp::path("detailed"))
-        // .and(warp::path::param::<u32>())
-        // .with(warp::compression::gzip())
-        .and(warp::body::content_length_limit(1024 * 16))
-        .and(warp::body::json())
-        .map(|request: request::DetailedRequest| {
-            warp::reply::json(&request.convert_to_internal_problem())
-        });
-
-    let routes = trip
-        .or(forward_geocoding).or(reverse_geocoding);
-
-    warp::serve(routes)
-        .run(([127, 0, 0, 1], 3030))
-        .await;
-}
-
-async fn receive_and_search_coordinates(postcode: String) -> Result<impl warp::Reply, Infallible> {
-    let result = geocoding::search_coordinates(postcode.as_ref());
-    Ok(result)
-}
-
-async fn receive_and_search_postcode(lat: f64, lon: f64) -> Result<impl warp::Reply, Infallible> {
-    let result = geocoding::search_postcode(vec![lat, lon]);
-    Ok(result)
+    let socket = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 3030);
+    start_server(socket).await;
 }
