+++ tests/geocoding.rs
@@ -1,6 +1,1 @@
-use crate:s:geocoding;
-
-#[test]
-fn test_geocoding_forward() {
-    assert_eq!(geocoding::search_coordinates, 0)
-}
+extern crate grekko;
